Refactor repositories with solid principles
