+++ OnlinerTask/OnlinerTask.Data/Repository/IProductRepository.cs
@@ -1,17 +1,13 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using OnlinerTask.Data.DataBaseModels;
+using OnlinerTask.Data.ScheduleModels;
 using OnlinerTask.Data.SearchModels;
 
 namespace OnlinerTask.Data.Repository
 {
     public interface IProductRepository
     {
-        Task<bool> CheckItem(int itemId, string username);
         Task<List<ProductModel>> CheckProducts(List<ProductModel> products, string userName);
-        List<Product> GetPersonalProducts(string name);
-        List<Product> GetAllProducts();
-        Task<string> RemoveOnlinerProduct(int itemId, string name);
-        bool CreateOnlinerProduct(ProductModel model, string userEmail);
+        UsersUpdateEmail WriteUpdate(ProductModel item, string useremail);
     }
 }
