+++ OnlinerTask/OnlinerTask.Data/Repository/ITimeServiceRepository.cs
@@ -1,16 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using OnlinerTask.Data.DataBaseModels;
-using OnlinerTask.Data.ScheduleModels;
-using OnlinerTask.Data.SearchModels;
+﻿using System.Threading.Tasks;
+using OnlinerTask.Data.Requests;
 
 namespace OnlinerTask.Data.Repository
 {
     public interface ITimeServiceRepository
     {
-        bool UpdateProduct(Product item, ProductModel model);
-        UsersUpdateEmail WriteUpdateToProduct(ProductModel item, TimeSpan time);
-        UsersUpdateEmail WriteUpdate(ProductModel item, string useremail);
-        List<Product> GetAllProducts();
+        Task ChangeSendEmailTimeAsync(TimeRequest request, string userName);
     }
 }
