+++ OnlinerTask/OnlinerTask.Data/Repository/Implementations/MsSQLProductRepository.cs
@@ -1,7 +1,11 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Data.Entity;
+using System.Linq;
 using System.Threading.Tasks;
 using OnlinerTask.Data.DataBaseModels;
+using OnlinerTask.Data.IdentityModels;
+using OnlinerTask.Data.ScheduleModels;
 using OnlinerTask.Data.SearchModels;
 using OnlinerTask.Extensions.Extensions;
 
@@ -9,13 +13,6 @@ namespace OnlinerTask.Data.Repository.Implementations
 {
     public class MsSqlProductRepository : IProductRepository
     {
-        private readonly IRepository repository;
-
-        public MsSqlProductRepository(IRepository repository)
-        {
-            this.repository = repository;
-        }
-
         public async Task<List<ProductModel>> CheckProducts(List<ProductModel> products, string userName)
         {
             await products.ForEachAsync(async i =>
@@ -25,7 +22,20 @@ namespace OnlinerTask.Data.Repository.Implementations
             return products;
         }
 
-        public async Task<bool> CheckItem(int itemId, string username)
+        public UsersUpdateEmail WriteUpdate(ProductModel item, string useremail)
+        {
+            using (var db = new ApplicationDbContext())
+            {
+                var time = db.Users.FirstOrDefault(x => x.Email == useremail);
+                if (time != null)
+                {
+                    return WriteUpdateToProduct(item, time.EmailTime);
+                }
+            }
+            return null;
+        }
+
+        private async Task<bool> CheckItem(int itemId, string username)
         {
             using (var context = new OnlinerProducts())
             {
@@ -33,24 +43,42 @@ namespace OnlinerTask.Data.Repository.Implementations
             }
         }
 
-        public List<Product> GetAllProducts()
+        private void UpdateProduct(Product item, ProductModel model)
         {
-            return repository.GetAllProducts();
+            if (item == null)
+            {
+                return;
+            }
+            using (var db = new OnlinerProducts())
+            {
+                var product = db.Product.FirstOrDefault(x => x.ProductId == item.ProductId);
+                if (product == null)
+                {
+                    return;
+                }
+                product.Price.PriceMaxAmmount.Amount = model.Prices.PriceMax.Amount;
+                product.Price.PriceMinAmmount.Amount = model.Prices.PriceMin.Amount;
+                db.SaveChanges();
+            }
         }
 
-        public List<Product> GetPersonalProducts(string name)
+        private UsersUpdateEmail WriteUpdateToProduct(ProductModel model, TimeSpan time)
         {
-            return repository.GetPersonalProducts(name);
-        }
-
-        public Task<string> RemoveOnlinerProduct(int itemId, string name)
-        {
-            return repository.RemoveOnlinerProduct(itemId, name);
-        }
-
-        public bool CreateOnlinerProduct(ProductModel model, string userEmail)
-        {
-            return repository.CreateOnlinerProduct(model, userEmail);
+            if (model == null)
+            {
+                return null;
+            }
+            using (var db = new OnlinerProducts())
+            {
+                var dbmodel = db.Product.FirstOrDefault(x => x.ProductId == model.Id);
+                UpdateProduct(dbmodel, model);
+                return new UsersUpdateEmail()
+                {
+                    ProductName = model.FullName,
+                    UserEmail = dbmodel.UserEmail,
+                    Time = time
+                };
+            }
         }
     }
 }
