+++ OnlinerTask/OnlinerTask.Data/Repository/Implementations/MsSQLTimeServiceRepository.cs
@@ -1,10 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Data.Entity;
+using System.Threading.Tasks;
 using OnlinerTask.Data.DataBaseModels;
 using OnlinerTask.Data.EntityMappers;
 using OnlinerTask.Data.IdentityModels;
-using OnlinerTask.Data.ScheduleModels;
+using OnlinerTask.Data.Requests;
 using OnlinerTask.Data.SearchModels;
 
 namespace OnlinerTask.Data.Repository.Implementations
@@ -18,61 +17,17 @@ namespace OnlinerTask.Data.Repository.Implementations
             this.repository = repository;
         }
 
-        public bool UpdateProduct(Product item, ProductModel model)
-        {
-            if (item == null)
-            {
-                return false;
-            }
-            using (var db = new OnlinerProducts())
-            {
-                var product = db.Product.FirstOrDefault(x => x.ProductId == item.ProductId);
-                if (product == null)
-                {
-                    return false;
-                }
-                product.Price.PriceMaxAmmount.Amount = model.Prices.PriceMax.Amount;
-                product.Price.PriceMinAmmount.Amount = model.Prices.PriceMin.Amount;
-                db.SaveChanges();
-                return true;
-            }
-        }
-
-        public UsersUpdateEmail WriteUpdateToProduct(ProductModel model, TimeSpan time)
-        {
-            if (model == null)
-            {
-                return null;
-            }
-            using (var db = new OnlinerProducts())
-            {
-                var dbmodel = db.Product.FirstOrDefault(x => x.ProductId == model.Id);
-                UpdateProduct(dbmodel, model);
-                return new UsersUpdateEmail()
-                {
-                    ProductName = model.FullName,
-                    UserEmail = dbmodel.UserEmail,
-                    Time = time
-                };
-            }
-        }
-
-        public UsersUpdateEmail WriteUpdate(ProductModel item, string useremail)
+        public async Task ChangeSendEmailTimeAsync(TimeRequest request, string userName)
         {
             using (var db = new ApplicationDbContext())
             {
-                var time = db.Users.FirstOrDefault(x => x.Email == useremail);
-                if (time != null)
+                var user = await db.Users.FirstOrDefaultAsync(x => x.UserName == userName);
+                if (user != null && request != null)
                 {
-                    return WriteUpdateToProduct(item, time.EmailTime);
+                    user.EmailTime = request.Time.TimeOfDay;
+                    await db.SaveChangesAsync();
                 }
             }
-            return null;
-        }
-
-        public List<Product> GetAllProducts()
-        {
-            return repository.GetAllProducts();
         }
     }
 }
